+++ lib/presentation/home_screen/bloc/movie_now_playing/movie_now_playing_bloc.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:ilia_movies/apis/now_playing_movies/domain/usecases/movie_usecase.dart';
+import 'package:ilia_movies/apis/now_playing_movies/domain/usecases/movie_now_playing_usecase.dart';
 import 'movie_now_playing_event.dart';
 import 'movie_now_playing_state.dart';
