+++ test/movie_usecase_test.dart
@@ -1,6 +1,6 @@
 import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:ilia_movies/apis/now_playing_movies/domain/usecases/movie_usecase.dart';
+import 'package:ilia_movies/apis/now_playing_movies/domain/usecases/movie_now_playing_usecase.dart';
 import 'package:ilia_movies/models/movie_entity.dart';
 import 'package:ilia_movies/presentation/home_screen/bloc/movie_now_playing/movie_now_playing_bloc.dart';
 import 'package:ilia_movies/presentation/home_screen/bloc/movie_now_playing/movie_now_playing_event.dart';
